FrontEnd/FUD/app/src/main/java/com/example/fud/Controller/AppController.java
package com.example.fud.Controller;
import android.content.Context;
import android.graphics.Bitmap;
import android.text.TextUtils;
import android.util.LruCache;

import com.android.volley.Request;
import com.android.volley.RequestQueue;
import com.android.volley.toolbox.ImageLoader;
import com.android.volley.toolbox.Volley;

/**
 * This class is responsible for managing Android Volley requests
 */
public class AppController {

    private static Context ctx;
    private ImageLoader mImageLoader;
    private RequestQueue mRequestQueue;
    private static AppController mInstance;

    /**
     * Tag is the tag for Volley requests
     */
    public static final String TAG = AppController.class.getSimpleName();

    private AppController(Context context) {
        ctx = context;
        mRequestQueue = getRequestQueue();

        mImageLoader = new ImageLoader(mRequestQueue, new ImageLoader.ImageCache() {
            private final LruCache<String, Bitmap> cache = new LruCache<String, Bitmap>(20);

            @Override
            public Bitmap getBitmap(String url) {
                return cache.get(url);
            }

            @Override
            public void putBitmap(String url, Bitmap bitmap) {
                cache.put(url, bitmap);
            }
        });
    }

    /**
     * This function returns the current instance of the app controller
     *
     * @param context
     * @return mInstance
     */
    public static synchronized AppController getInstance(Context context) {
        if (mInstance == null) {
            mInstance = new AppController(context);
        }
        return mInstance;
    }

    /**
     * This function returns the current request queue
     *
     * @return mRequestQueue
     */
    public RequestQueue getRequestQueue() {
        if (mRequestQueue == null) {
            mRequestQueue = Volley.newRequestQueue(ctx.getApplicationContext());
        }
        return mRequestQueue;
    }

    /**
     * This method adds a request to the request queue with a specific tag
     *
     * @param req
     * @param tag
     * @param <T>
     */
    public <T> void addToRequestQueue(Request<T> req, String tag) {
        // set the default tag if tag is empty
        req.setTag(TextUtils.isEmpty(tag) ? TAG : tag);
        getRequestQueue().add(req);
    }

    /**
     * This method adds a request to the request queue
     *
     * @param req
     * @param <T>
     */
    public <T> void addToRequestQueue(Request<T> req) {
        req.setTag(TAG);
        getRequestQueue().add(req);
    }

    /**
     * This method cancels pending requests in the request queue
     *
     * @param tag
     */
    public void cancelPendingRequests(Object tag) {
        if (mRequestQueue != null) {
            mRequestQueue.cancelAll(tag);
        }
    }

    /**
     * Returns the current instance of the Image Loader
     *
     * @return mImageLoader
     */
    public ImageLoader getImageLoader() {
        return mImageLoader;
    }
}

